Extract test functions into test.h.
