+++ Core/Tests/Utils/LUT.c
@@ -1,56 +1,9 @@
 #include "../Inc/Utils/LUT.h"
+#include "test.h"
 
 #include <assert.h>
 #include <math.h>
 #include <stdio.h>
-
-#ifndef TEST_EQUAL_EPSILON
-#define TEST_EQUAL_EPSILON 0.1
-#endif
-
-#ifndef TEST_ERROR_DELTA_PERCENT
-#define TEST_ERROR_DELTA_PERCENT 5
-#endif
-
-// Prints a test start message.
-void test_start(const char *test_label) { printf("START\t%s\n", test_label); }
-
-// Prints a message depending on if the test passes or fails
-void test(const char *pass_message, const char *fail_message, bool passes) {
-  if (passes) {
-    printf("PASS\t%s\n", pass_message);
-  } else {
-    printf("FAIL\t%s\n", fail_message);
-  }
-}
-
-// Tests if the values are approximately equal (pass) or
-// not (fail); approximately equal depends on the value of ϵ
-// (TEST_EQUAL_EPSILON)
-void test_equal(const char *a_label, const char *b_label, float a, float b) {
-  bool passes = fabs(a - b) < TEST_EQUAL_EPSILON;
-  if (passes) {
-    printf("PASS\t%s = %s (%f = %f)\n", a_label, b_label, a, b);
-  } else {
-    printf("FAIL\t%s ≠ %s (%f ≠ %f)\n", a_label, b_label, a, b);
-  }
-}
-
-// Tests if the values are within a percent error threshold (pass) or not
-// (fail); percent error threshold depends on the value of δ
-// (TEST_ERROR_DELTA_PERCENT)
-void test_within_error(const char *actual_label, const char *expected_label,
-                       float actual, float expected) {
-  float percent_error = (actual - expected) / expected * 100;
-  bool passes = fabs(percent_error) < TEST_ERROR_DELTA_PERCENT;
-  if (passes) {
-    printf("PASS\t%s ≈ %s (%f ≈ %f)\n", actual_label, expected_label, actual,
-           expected);
-  } else {
-    printf("FAIL\t%s ≠ %s (%f ≠ %f)\n", actual_label, expected_label, actual,
-           expected);
-  }
-}
 
 // Generates expected altitude measurements in meters for pressure readings in
 // millibar. Uses the international barometric formula.
